pyproject.toml
[project]
name = "data-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "openai>=2.21.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
